backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
python-multipart==0.0.12
pymongo==4.6.0
motor==3.3.2
groq==0.37.1
certifi==2025.11.12
firebase-admin==6.6.0
pydantic==2.10.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
PyPDF2==3.0.1
python-docx==1.1.2
pytesseract==0.3.10
Pillow==11.0.0
pdf2image==1.17.0
python-magic-bin==0.4.14
